Fix method for updating project responsible person

+++ UJCSystem/Vokabular.MainService.DataContracts/Contracts/ResponsiblePersonContract.cs
@@ -14,7 +14,7 @@
 
     public class ProjectResponsiblePersonIdContract
     {
-        public long ResponsiblePersonId { get; set; }
-        public long ResponsibleTypeId { get; set; }
+        public int ResponsiblePersonId { get; set; }
+        public int ResponsibleTypeId { get; set; }
     }
 }
